Add more robust exception handling and controller action

+++ src/main/java/com/scar/lms/service/impl/BorrowServiceImpl.java
@@ -1,8 +1,10 @@
 package com.scar.lms.service.impl;
 
 import com.scar.lms.entity.Borrow;
+import com.scar.lms.exception.OperationNotAllowedException;
 import com.scar.lms.repository.BorrowRepository;
 import com.scar.lms.service.BorrowService;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,6 +29,9 @@ public class BorrowServiceImpl implements BorrowService {
 
     @Override
     public void addBorrow(Borrow borrow) {
+        if (borrowRepository.existsById(borrow.getId())) {
+            throw new OperationNotAllowedException("Unable to borrow book");
+        }
         borrowRepository.save(borrow);
     }
 
@@ -37,6 +42,9 @@ public class BorrowServiceImpl implements BorrowService {
 
     @Override
     public void removeBorrow(Borrow borrow) {
+        if (!borrowRepository.existsById(borrow.getId())) {
+            throw new OperationNotAllowedException("Unable to remove book");
+        }
         borrowRepository.delete(borrow);
     }
